# szteam-network-frontend/src/app/services/models/review-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ReviewResponse {
  author?: string;
  comment?: string;
  createdAt?: string;
  ownReview?: boolean;
  score?: number;
}