# szteam-network-frontend/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenApi Specifikáció - Kristóf",
    "description": "OpenApi dokumentáció a Spring securityhez",
    "termsOfService": "Terms of Service",
    "contact": {
      "name": "Kristóf",
      "email": "kristof.zsoldos@gmail.com"
    },
    "license": {
      "name": "Licenese név"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8088/api/v1",
      "description": "LOCAL ENV"
    },
    {
      "url": "http://fakeurlfornow.com",
      "description": "PROD ENV"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/reviews": {
      "post": {
        "tags": [
          "Review"
        ],
        "operationId": "saveReview",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
...
    "/reviews/game/{game-id}": {
      "get": {
        "tags": [
          "Review"
        ],
        "operationId": "findAllReviewsByGame",
        "parameters": [
          {
            "name": "game-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseReviewResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ReviewRequest": {
        "required": [
          "comment",
          "gameId"
        ],
        "type": "object",
        "properties": {
          "score": {
            "maximum": 5,
            "minimum": 0,
            "type": "number",
            "format": "double"
          },
          "comment": {
            "type": "string"
          },
          "gameId": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "PageResponseReviewResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
      "ReviewResponse": {
        "type": "object",
        "properties": {
          "score": {
            "type": "number",
            "format": "double"
          },
          "comment": {
            "type": "string"
          },
          "ownReview": {
            "type": "boolean"
          },
          "author": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT auth leírás",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}